+++ cypress/integration/v2/blocks/testimonial/mobile-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { mobileStyle } from '.'
+
+describe( 'Testimonial Block ( Mobile Style )', registerTests( [ mobileStyle ] ) )
