+++ cypress/integration/v2/blocks/testimonial/other-tests.spec.js
@@ -1,0 +1,23 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import {
+	blockExist,
+	blockError,
+	switchLayout,
+	switchDesign,
+	typeContent,
+} from '.'
+
+describe( 'Testimonial Block ( Other Tests )', registerTests( [
+	blockExist,
+	blockError,
+	switchLayout,
+	switchDesign,
+	typeContent,
+] ) )
